refactor(e2e): extract request-queue helpers to test-dapp flow
Move shared dapp and network manager helpers out of ui.spec.ts into
test-dapp.flow.ts so they can be reused, and fix openNetworkAndSelectNetwork
to select networks by CAIP chain ID or display name before closing the modal.
